more transformations segregated
