+++ source/electrical_heating/src/opengeh_electrical_heating/domain/transformations/__init__.py
@@ -1,4 +1,10 @@
 from .common import calculate_daily_quantity
-from .consumption_energy import get_daily_consumption_energy_in_local_time
+from .metering_points import get_joined_metering_point_periods_in_local_time
+from .time_series_points import get_daily_consumption_energy_in_local_time, get_electrical_heating_in_local_time
 
-__all__ = ["calculate_daily_quantity", "get_daily_consumption_energy_in_local_time"]
+__all__ = [
+    "calculate_daily_quantity",
+    "get_joined_metering_point_periods_in_local_time",
+    "get_electrical_heating_in_local_time",
+    "get_daily_consumption_energy_in_local_time",
+]
